rpa/requirements.txt
selenium
webdriver-manager
supabase
python-decouple
fastapi
uvicorn
psycopg2-binary
